infra: remove control_output_node
Remove all references to control output node and the associated code.
This control_output node is no longer required as all features are
implemented as nodes.
