+++ modules/infra/control/gr_loopback.h
@@ -7,6 +7,5 @@
 
 #include <rte_byteorder.h>
 
-void loopback_tx(struct rte_mbuf *m);
 control_input_t loopback_get_control_id(void);
 void loopback_input_add_type(rte_be16_t eth_type, const char *next_node);
