+++ modules/infra/control/loopback.h
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: BSD-3-Clause
+// Copyright (c) 2025 Christophe Fontaine
+
+#pragma once
+
+struct iface_info_loopback {
+	int fd;
+	struct event *ev;
+};
