+++ modules/infra/datapath/meson.build
@@ -3,13 +3,11 @@
 
 src += files(
   'control_input.c',
-  'control_output.c',
   'drop.c',
   'eth_input.c',
   'eth_output.c',
   'l1_xconnect.c',
   'loop_input.c',
-  'loop_output.c',
   'main_loop.c',
   'rx.c',
   'trace.c',
