+++ modules/ip/datapath/ip_output.c
@@ -1,7 +1,6 @@
 // SPDX-License-Identifier: BSD-3-Clause
 // Copyright (c) 2024 Robin Jarry
 
-#include <gr_control_output.h>
 #include <gr_datapath.h>
 #include <gr_eth.h>
 #include <gr_fib4.h>
